# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server for a recipe API.",
        "title": "Recipe API",
        "contact": {
            "name": "Milan Poudel",
            "url": "https://github.com/milanpoudelwebdeveloper"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/admin/users": {
            "get": {
                "description": "Get all users",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Get all users",
                "responses": {
                    "200": {
                        "description": "List of users",
                        "schema": {
                            "$ref": "#/definitions/models.User"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Create a new user",
                "parameters": [
                    {
                        "description": "New User",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/admin.NewUser"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "New user created successfully",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/recipes": {
            "get": {
                "description": "get all recipes",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "recipes"
                ],
                "summary": "get all recipes",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Recipe"
                        }
                    }
                }
            },
            "post": {
                "description": "Post a recipe",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "recipes"
                ],
                "summary": "Post a recipe",
                "parameters": [
                    {
                        "description": "Recipe object that needs to be added",
                        "name": "recipe",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Recipe"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Recipe added successfully",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/recipes/{id}": {
            "get": {
                "description": "Get recipe details",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "recipes"
                ],
                "summary": "Get recipe details",
                "responses": {
                    "200": {
                        "description": "Get recipe details",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "put": {
                "description": "Update recipe",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "recipes"
                ],
                "summary": "Update recipe",
                "parameters": [
                    {
                        "description": "Recipe object that needs to be updated",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Recipe"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Update recipe",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete recipe",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "recipes"
                ],
                "summary": "Delete recipe",
                "responses": {
                    "200": {
                        "description": "Delete recipe",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "admin.NewUser": {
            "type": "object",
            "required": [
                "email",
                "role"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                }
            }
        },
...
        "models.Recipe": {
            "type": "object",
            "required": [
                "country",
                "description",
                "id",
                "instructions",
                "title"
            ],
            "properties": {
                "country": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "instructions": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "models.User": {
            "type": "object",
            "required": [
                "country",
                "email",
                "id",
                "name",
                "password",
                "profile_image",
                "role",
                "verified"
            ],
            "properties": {
                "country": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "profile_image": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                },
                "verified": {
                    "type": "boolean"
                }
            }
        }
    }
}